feat: add reconfigure script and make job
Adds reconfigure script for Dqlite. It will read info.yaml in the
default directory for Juju's dqlite data, and reconfigure the cluster as
a single node, according to its contents.

The make job builds the script as a stand-alone utility and leaves it in
the standard build directory (./_build/_/bin/reconfigure).